requirements.txt
streamlit>=1.27.0
streamlit-lottie>=0.0.5
streamlit-timeline>=0.0.2
llama-index-core>=0.10.0
llama-index-llms-openai>=0.1.0
openai>=1.0.0
pillow>=9.0.0
langchain>=0.0.267
requests>=2.31.0